pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = ""
authors = ["Samuel Agyei <88740526+SterotECH@users.noreply.github.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-unfold = "^0.6.5"
django-filter = "^23.2"
django-markdown = "^0.8.4"
python-dotenv = "^1.0.0"
django-cors-headers = "^4.1.0"
django-mail-templated = "^2.6.5"
daphne = "^4.0.0"
channels = "^4.0.0"
psycopg2 = "^2.9.6"
celery = "^5.3.4"
redis = "^5.0.0"
django-redis = "^5.3.0"
whitenoise = "^6.5.0"
django-celery-beat = "^2.5.0"
gunicorn = "^21.2.0"
django-import-export = "^3.3.1"
django-environ = "^0.11.2"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[tool.ruff]
fix = true
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    "**/migrations"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
